Move the new indent code to its own folder within rules
