+++ packages/eslint-plugin/src/rules/indent-new-do-not-use/TokenInfo.ts
