--- packages/eslint-plugin/src/util/indent/index.ts
@@ -1,3 +1,0 @@
-export * from './BinarySearchTree';
-export * from './OffsetStorage';
-export * from './TokenInfo';
